+++ src/main/java/com/shoppingcartservice/sachin/Entities/Product/Subcategory.java
@@ -1,18 +1,13 @@
 package com.shoppingcartservice.sachin.Entities.Product;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
 
 @Entity
 public class Subcategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int subcategoryId;
+    private long subcategoryId;
     private String name;
-
-
 
     public Subcategory() {
     }
@@ -21,11 +16,11 @@ public class Subcategory {
         this.name = name;
     }
 
-    public int getSubcategoryId() {
+    public long getSubcategoryId() {
         return subcategoryId;
     }
 
-    public void setSubcategoryId(int subcategoryId) {
+    public void setSubcategoryId(long subcategoryId) {
         this.subcategoryId = subcategoryId;
     }
 
@@ -36,6 +31,4 @@ public class Subcategory {
     public void setName(String name) {
         this.name = name;
     }
-
-
 }
