+++ src/main/java/com/shoppingcartservice/sachin/Reposistories/User/AddressRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressRepo extends JpaRepository<Address,Integer> {
-    Address findAddressByAddressID(Integer addressId);
+public interface AddressRepo extends JpaRepository<Address,Long> {
+    Address findAddressByAddressID(Long addressId);
 }
